chore: spilt code to hooks
