+++ frontend/webapp/hooks/common/useContainerWidth.ts
@@ -1,0 +1,23 @@
+import { useEffect, useState, useRef } from 'react';
+
+export function useContainerWidth() {
+  const containerRef = useRef<HTMLDivElement | null>(null);
+  const [containerWidth, setContainerWidth] = useState<number>(0);
+
+  useEffect(() => {
+    const updateWidth = () => {
+      if (containerRef.current) {
+        setContainerWidth(
+          containerRef.current.getBoundingClientRect().width - 64
+        );
+      }
+    };
+
+    updateWidth();
+
+    window.addEventListener('resize', updateWidth);
+    return () => window.removeEventListener('resize', updateWidth);
+  }, []);
+
+  return { containerRef, containerWidth };
+}
