+++ frontend/webapp/hooks/index.tsx
@@ -10,3 +10,5 @@ export * from './useSSE';
 export * from './new-config';
 export * from './compute-platform';
 export * from './useOverviewMetrics';
+export * from './overview';
+export * from './common';
